http_to_mqtt_api/requirements.txt
paho-mqtt
fastapi
uvicorn